+++ src/php/Game/Action/Exception/MoveOverNumberOfUnitsException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Hexammon\HexoNards\Game\Action\Exception;
+
+use Hexammon\HexoNards\Exception\DomainException;
+
+class MoveOverNumberOfUnitsException extends DomainException
+{
+
+}