+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased] 
 ### Fixed
 - One unit army moving: bug with zero-size army after move.
+- Check number of units in MoveArmy action.
 
 ## 1.0.0-alpha.4 - 2017-11-29
 ### Added